refactor : merge header parser and body param parser
