+++ src/main/java/webserver/RequestLine.java
@@ -1,0 +1,4 @@
+package webserver;
+
+public class RequestLine {
+}
